Cambio error JpaRepository

+++ src/main/java/com/pfc/octobets/repository/dao/ApuestaRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.pfc.octobets.repository.entity.Apuesta;
 
-public interface ApuestaRepository extends JpaRepository<Apuesta, Integer> {
+public interface ApuestaRepository extends JpaRepository<Apuesta, Long> {
 
 }
