+++ src/main/java/com/pfc/octobets/repository/dao/CarteraRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.pfc.octobets.repository.entity.Cartera;
 
-public interface CarteraRepository extends JpaRepository<Cartera, Integer> {
+public interface CarteraRepository extends JpaRepository<Cartera, Long> {
 
 }
