+++ src/main/java/com/pfc/octobets/repository/dao/JuegoRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.pfc.octobets.repository.entity.Juego;
 
-public interface JuegoRepository extends JpaRepository<Juego, Integer> {
+public interface JuegoRepository extends JpaRepository<Juego, Long> {
 
 }
