+++ src/main/java/com/pfc/octobets/repository/dao/OpcionRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.pfc.octobets.repository.entity.Opcion;
 
-public interface OpcionRepository extends JpaRepository<Opcion, Integer> {
+public interface OpcionRepository extends JpaRepository<Opcion, Long> {
 
 }
