+++ src/main/java/com/pfc/octobets/repository/dao/RondaRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.pfc.octobets.repository.entity.Ronda;
 
-public interface RondaRepository extends JpaRepository<Ronda, Integer> {
+public interface RondaRepository extends JpaRepository<Ronda, Long> {
 
 }
