+++ src/main/java/com/pfc/octobets/repository/dao/TransaccionRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.pfc.octobets.repository.entity.Transaccion;
 
-public interface TransaccionRepository extends JpaRepository<Transaccion, Integer> {
+public interface TransaccionRepository extends JpaRepository<Transaccion, Long> {
 
 }
